Move Requests to root folder and make package private
