+++ src/main/java/com/treescrub/spedran/HttpRequestException.java
@@ -1,4 +1,4 @@
-package com.treescrub.spedran.requests;
+package com.treescrub.spedran;
 
 import kong.unirest.HttpRequest;
 import kong.unirest.HttpResponse;
