+++ src/main/java/com/treescrub/spedran/ResourceRequest.java
@@ -1,4 +1,4 @@
-package com.treescrub.spedran.requests;
+package com.treescrub.spedran;
 
 import kong.unirest.HttpMethod;
 import kong.unirest.HttpRequest;
