+++ src/main/java/com/treescrub/spedran/Spedran.java
@@ -1,6 +1,5 @@
 package com.treescrub.spedran;
 
-import com.treescrub.spedran.requests.Requests;
 import com.treescrub.spedran.requests.builders.*;
 import com.treescrub.spedran.requests.builders.run.*;
 import com.treescrub.spedran.requests.builders.series.AllSeriesRequest;
