+++ src/main/java/com/treescrub/spedran/requests/builders/GametypesRequest.java
@@ -1,7 +1,7 @@
 package com.treescrub.spedran.requests.builders;
 
 import com.treescrub.spedran.data.Gametype;
-import com.treescrub.spedran.requests.ResourceCollectionRequest;
+import com.treescrub.spedran.ResourceCollectionRequest;
 import com.treescrub.spedran.requests.SortDirection;
 import kong.unirest.HttpMethod;
