+++ src/main/java/com/treescrub/spedran/requests/builders/LeaderboardRequest.java
@@ -1,6 +1,6 @@
 package com.treescrub.spedran.requests.builders;
 
-import com.treescrub.spedran.requests.SingleResourceRequest;
+import com.treescrub.spedran.SingleResourceRequest;
 import com.treescrub.spedran.data.*;
 import kong.unirest.HttpMethod;
