+++ src/main/java/com/treescrub/spedran/requests/builders/NotificationsRequest.java
@@ -2,7 +2,7 @@ package com.treescrub.spedran.requests.builders;
 
 import com.treescrub.spedran.data.Notification;
 import com.treescrub.spedran.data.User;
-import com.treescrub.spedran.requests.ResourceCollectionRequest;
+import com.treescrub.spedran.ResourceCollectionRequest;
 import kong.unirest.HttpMethod;
 
 /**
