+++ src/main/java/com/treescrub/spedran/requests/builders/PublisherRequest.java
@@ -1,7 +1,7 @@
 package com.treescrub.spedran.requests.builders;
 
 import com.treescrub.spedran.data.Publisher;
-import com.treescrub.spedran.requests.SingleResourceRequest;
+import com.treescrub.spedran.SingleResourceRequest;
 import kong.unirest.HttpMethod;
 
 import java.util.Map;
