+++ src/main/java/com/treescrub/spedran/requests/builders/PublishersRequest.java
@@ -2,7 +2,7 @@ package com.treescrub.spedran.requests.builders;
 
 import com.treescrub.spedran.data.Gametype;
 import com.treescrub.spedran.data.Publisher;
-import com.treescrub.spedran.requests.ResourceCollectionRequest;
+import com.treescrub.spedran.ResourceCollectionRequest;
 import com.treescrub.spedran.requests.SortDirection;
 import kong.unirest.HttpMethod;
