+++ src/main/java/com/treescrub/spedran/requests/builders/run/RunRequest.java
@@ -1,7 +1,7 @@
 package com.treescrub.spedran.requests.builders.run;
 
 import com.treescrub.spedran.data.Run;
-import com.treescrub.spedran.requests.SingleResourceRequest;
+import com.treescrub.spedran.SingleResourceRequest;
 import kong.unirest.HttpMethod;
 
 import java.util.Map;
