+++ src/main/java/com/treescrub/spedran/requests/builders/run/RunStatusRequest.java
@@ -3,7 +3,7 @@ package com.treescrub.spedran.requests.builders.run;
 import com.treescrub.spedran.data.Run;
 import com.treescrub.spedran.data.SubmissionStatus;
 import com.treescrub.spedran.requests.InvalidBuilderStateException;
-import com.treescrub.spedran.requests.ModifyResourceRequest;
+import com.treescrub.spedran.ModifyResourceRequest;
 import kong.unirest.HttpMethod;
 import kong.unirest.json.JSONElement;
 import kong.unirest.json.JSONObject;
