+++ src/main/java/com/treescrub/spedran/requests/builders/series/AllSeriesRequest.java
@@ -2,7 +2,7 @@ package com.treescrub.spedran.requests.builders.series;
 
 import com.treescrub.spedran.data.Series;
 import com.treescrub.spedran.data.User;
-import com.treescrub.spedran.requests.ResourceCollectionRequest;
+import com.treescrub.spedran.ResourceCollectionRequest;
 import com.treescrub.spedran.requests.SortDirection;
 import kong.unirest.HttpMethod;
