+++ src/main/java/com/treescrub/spedran/requests/builders/user/UsersRequest.java
@@ -1,7 +1,7 @@
 package com.treescrub.spedran.requests.builders.user;
 
 import com.treescrub.spedran.data.User;
-import com.treescrub.spedran.requests.ResourceCollectionRequest;
+import com.treescrub.spedran.ResourceCollectionRequest;
 import com.treescrub.spedran.requests.SortDirection;
 import kong.unirest.HttpMethod;
